runtime as shared app state
